Tarea Final

+++ libreria/example.h
@@ -3,5 +3,5 @@
 #include <iostream>
 using namespace std;
 
-float calcularIMC(int peso, int altura);
+float calcularIMC(int peso, float altura);
 void resolvente(int a, int b, int c);
